Time: 75 ms (6.35%), Space: 42.2 MB (20.63%) - LeetHub

+++ 0535-encode-and-decode-tinyurl/0535-encode-and-decode-tinyurl.cs
@@ -1,7 +1,7 @@
 public class Codec {
 
-    private static Dictionary<int, string> idToUrl = new Dictionary<int, string>();
-    private static readonly Dictionary<int, char> base16 = new Dictionary<int, char> {
+    private static Dictionary<long, string> idToUrl = new Dictionary<long, string>();
+    private static readonly Dictionary<long, char> base16 = new Dictionary<long, char> {
         {0, '0'},
         {1, '1'},
         {2, '2'},
@@ -20,7 +20,7 @@ public class Codec {
         {15, 'F'},
     };
 
-    private static readonly Dictionary<char, int> base16Rev = new Dictionary<char, int> {
+    private static readonly Dictionary<char, long> base16Rev = new Dictionary<char, long> {
         {'A', 10},
         {'B', 11},
         {'C', 12},
@@ -29,7 +29,7 @@ public class Codec {
         {'F', 15},
     };
 
-    private static int id = 1;
+    private static long id = 1;
 
     // Encodes a URL to a shortened URL
     public string encode(string longUrl) {
@@ -41,11 +41,11 @@ public class Codec {
     // Decodes a shortened URL to its original URL.
     public string decode(string shortUrl) {
         string[] parts = shortUrl.Split("/");
-        int decoded = convertToBase10(parts[parts.Length - 1]);
+        long decoded = convertToBase10(parts[parts.Length - 1]);
         return idToUrl[decoded];
     }
 
-    private string convertTo16Base(int i) {
+    private string convertTo16Base(long i) {
         var sb = new StringBuilder();
         while (i > 0) {
             sb.Append(base16[i % 16].ToString());
@@ -54,8 +54,8 @@ public class Codec {
         return sb.ToString();
     }
 
-    private int convertToBase10(string s) {
-        int res = 0;
+    private long convertToBase10(string s) {
+        long res = 0;
         foreach (char c in s) {
             if (base16Rev.ContainsKey(c)) res += res * 16 + base16Rev[c];
             else res += res * 16 + int.Parse(c.ToString());
